+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## Gold & Currency Tracker
-### [1.1.1](https://github.com/diomsg-code/GoldCurrencyTracker/tree/1.1.1) (@build-date@)
-[Full Changelog](https://github.com/diomsg-code/GoldCurrencyTracker/compare/1.1.0...1.1.1) - [Previous Releases](https://github.com/diomsg-code/GoldCurrencyTracker/releases)
+### [1.2.0](https://github.com/diomsg-code/GoldCurrencyTracker/tree/1.2.0) (@build-date@)
+[Full Changelog](https://github.com/diomsg-code/GoldCurrencyTracker/compare/1.1.1...1.2.0) - [Previous Releases](https://github.com/diomsg-code/GoldCurrencyTracker/releases)
 
-- Fixed: The overview window also opened automatically after a zone change
+- Changed: Separation of character and warband currencies in the selection dropdown
+- Changed: New database structure, the previous data is migrated